% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamical-systems-notes-linear-constructor.R
\name{LinearConstructor}
\alias{LinearConstructor}
\title{Simple Linear System Function Constructor}
\usage{
LinearConstructor(alpha, beta, sigmasq = NULL)
}
\arguments{
\item{alpha}{Numeric.
Intercept \eqn{\alpha}.}

\item{beta}{Numeric.
Slope \eqn{\beta}.}

\item{sigmasq}{Positive numeric value.
Variance of the random error term with expected value of zero.
If \code{sigmasq = NULL}, the system is deterministic.}
}
\description{
Simple Linear System Function Constructor
}
\details{
The simple linear system is given by
\deqn{
    y_{i t} = \alpha + \beta y_{i \left( t - 1 \right)}
  }
where
\eqn{\alpha} is the intercept and \eqn{\beta} is the slope.
}
\examples{
func <- LinearConstructor(alpha = 8.0, beta = 0.8)
func(0.1)

}
\author{
Ivan Jacob Agaloos Pesigan
}
