% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamical-systems-notes-logistic-constructor.R
\name{LogisticConstructor}
\alias{LogisticConstructor}
\title{Simple Logistic System Function Constructor}
\usage{
LogisticConstructor(r, K, sigmasq = NULL)
}
\arguments{
\item{r}{Numeric.
Growth rate.}

\item{K}{Numeric
Carrying capacity.}

\item{sigmasq}{Positive numeric value.
Variance of the random error term with expected value of zero.
If \code{sigmasq = NULL}, the system is deterministic.}
}
\description{
Simple Logistic System Function Constructor
}
\details{
The logistic system is given by
\deqn{
    y_{i t} = r y_{i \left( t - 1 \right)}
    \left( 1 - \frac{y_{i \left( t - 1 \right)}}{K} \right)
  }
where
\eqn{r} is the growth rate and \eqn{K} is the carrying capacity.
}
\examples{
func <- LogisticConstructor(r = 1.5, K = 10)
func(0.1)

}
\author{
Ivan Jacob Agaloos Pesigan
}
