% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamical-systems-notes-univariate-series.R
\name{UnivSeries}
\alias{UnivSeries}
\title{Univariate Time Series}
\usage{
UnivSeries(y0, func, nt)
}
\arguments{
\item{y0}{Numeric.
Initial condition.}

\item{func}{Function.
The input is \eqn{y_{i \left( t - 1\right)}}
and the output is \eqn{y_{i t}}.}

\item{nt}{Positive integer.
Maximum discrete time points.}
}
\description{
Univariate Time Series
}
\details{
The univariate time series
is generated using the following equation
\deqn{
    y_{i t} = f \left( y_{i \left( t - 1 \right) \right)} .
  }
}
\examples{
# linear
## deterministic
func <- LinearConstructor(alpha = 8.0, beta = 0.8)
y <- UnivSeries(y0 = 0.01, func = func, nt = 100)
plot(y)
## stochastic
func <- LinearConstructor(alpha = 8.0, beta = 0.8, sigmasq = 1)
y <- UnivSeries(y0 = 0.01, func = func, nt = 100)
plot(y)

# logistic
## deterministic
func <- LogisticConstructor(r = 1.5, K = 10)
y <- UnivSeries(y0 = 0.01, func = func, nt = 100)
plot(y)
## stochastic
func <- LogisticConstructor(r = 1.5, K = 10, sigmasq = 1)
y <- UnivSeries(y0 = 0.01, func = func, nt = 100)
plot(y)

}
\author{
Ivan Jacob Agaloos Pesigan
}
