% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamical-systems-notes-package.R
\docType{package}
\name{dynamicalSystemsNotes-package}
\alias{dynamicalSystemsNotes-package}
\alias{_PACKAGE}
\title{dynamicalSystemsNotes: Dynamical Systems Notes}
\description{
HDFS 538 Dynamical Systems Methods and Applications personal study notes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ijapesigan/dynamicalSystemsNotes}
  \item \url{https://ijapesigan.github.io/dynamicalSystemsNotes/}
  \item Report bugs at \url{https://github.com/ijapesigan/dynamicalSystemsNotes/issues}
}

}
\author{
\strong{Maintainer}: Ivan Jacob Agaloos Pesigan \email{ijapesigan@gmail.com} (\href{https://orcid.org/0000-0003-4818-8420}{ORCID}) [copyright holder]

}
\keyword{internal}
